+++ src/main/java/com/powsybl/sld/server/entities/nad/NadConfigEntity.java
@@ -1,4 +1,4 @@
-/*
+/**
  * Copyright (c) 2025, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -23,6 +23,7 @@ import java.util.*;
 @Builder
 @Table(name = "nadConfig")
 public class NadConfigEntity {
+
     public NadConfigEntity(NadConfigEntity origin) {
         this.depth = origin.getDepth();
         this.scalingFactor = origin.getScalingFactor();
@@ -53,7 +54,7 @@ public class NadConfigEntity {
     private Integer scalingFactor;
 
     @Column(name = "radiusFactor")
-    private Integer radiusFactor;
+    private Double radiusFactor;
 
     @OneToMany(cascade = CascadeType.ALL)
     @JoinColumn(name = "nad_config_id")
